requirements.txt
# RAG Starter Kit: Python Dependencies
# Python 3.12+ required (3.12 is supported until 2028-10)

# Core Dependencies
python-dotenv==1.0.1
streamlit==1.41.1
supabase==2.11.0
openai==1.59.6

# Async Support (required for Streamlit async operations)
nest-asyncio==1.6.0
